1032820: Moved implementations to Navigation Manager
