+++ src/Components/Components/src/NavigationManagerExtensions.cs
@@ -2,10 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Collections;
-using System.Collections.Concurrent;
 using System.Globalization;
 using System.Linq;
-using System.Reflection;
 using System.Text;
 using Microsoft.AspNetCore.Components.Routing;
 using Microsoft.AspNetCore.Internal;
@@ -40,8 +38,7 @@ public static class NavigationManagerExtensions
         [typeof(long)] = value => Format((long)value),
     };
 
-    // Cache for component route templates to avoid repeated reflection
-    private static readonly ConcurrentDictionary<Type, string[]> _componentRouteTemplates = new();
+    
 
     private static string? Format(string? value)
         => value;
@@ -837,126 +834,5 @@ public static class NavigationManagerExtensions
             hashValue.AsSpan().CopyTo(chars[position..]);
         });
     }
-
-    /// <summary>
-    /// Gets the URI for the specified component type with optional route parameters.
-    /// </summary>
-    /// <typeparam name="TComponent">The component type to navigate to.</typeparam>
-    /// <param name="navigationManager">The <see cref="NavigationManager"/>.</param>
-    /// <param name="parameters">The route parameters to include in the URI.</param>
-    /// <returns>The URI for the specified component.</returns>
-    /// <exception cref="InvalidOperationException">Thrown when the component type does not have a route attribute.</exception>
-    public static string GetUri<TComponent>(this NavigationManager navigationManager, params object[] parameters)
-    {
-        ArgumentNullException.ThrowIfNull(navigationManager);
-
-        var componentType = typeof(TComponent);
-        var routeTemplates = GetComponentRouteTemplates(componentType);
-
-        if (routeTemplates.Length == 0)
-        {
-            throw new InvalidOperationException($"Component '{componentType.FullName}' does not have a Route attribute.");
-        }
-
-        // For now, use the first route template (future enhancement: allow specifying which route to use)
-        var routeTemplate = routeTemplates[0];
-        return BindParametersToRouteTemplate(routeTemplate, parameters);
-    }
-
-    /// <summary>
-    /// Navigates to the specified component type with optional route parameters.
-    /// </summary>
-    /// <typeparam name="TComponent">The component type to navigate to.</typeparam>
-    /// <param name="navigationManager">The <see cref="NavigationManager"/>.</param>
-    /// <param name="parameters">The route parameters to include in the URI.</param>
-    /// <exception cref="InvalidOperationException">Thrown when the component type does not have a route attribute.</exception>
-    public static void NavigateTo<TComponent>(this NavigationManager navigationManager, params object[] parameters)
-    {
-        ArgumentNullException.ThrowIfNull(navigationManager);
-
-        var uri = navigationManager.GetUri<TComponent>(parameters);
-        navigationManager.NavigateTo(uri);
-    }
-
-    /// <summary>
-    /// Gets the route templates for the specified component type.
-    /// </summary>
-    /// <param name="componentType">The component type.</param>
-    /// <returns>An array of route templates.</returns>
-    private static string[] GetComponentRouteTemplates(Type componentType)
-    {
-        return _componentRouteTemplates.GetOrAdd(componentType, type =>
-        {
-            var routeAttributes = type.GetCustomAttributes<RouteAttribute>(inherit: false);
-            var routeAttributesArray = routeAttributes as RouteAttribute[] ?? routeAttributes.ToArray();
-            if (routeAttributesArray.Length == 0)
-            {
-                return Array.Empty<string>();
-            }
-
-            var templates = new string[routeAttributesArray.Length];
-            for (var i = 0; i < routeAttributesArray.Length; i++)
-            {
-                templates[i] = routeAttributesArray[i].Template;
-            }
-
-            return templates;
-        });
-    }
-
-    /// <summary>
-    /// Binds parameters to a route template.
-    /// </summary>
-    /// <param name="routeTemplate">The route template.</param>
-    /// <param name="parameters">The parameters to bind.</param>
-    /// <returns>The route template with parameters bound.</returns>
-    private static string BindParametersToRouteTemplate(string routeTemplate, object[] parameters)
-    {
-        if (parameters == null || parameters.Length == 0)
-        {
-            return routeTemplate;
-        }
-
-        var sb = new StringBuilder(capacity: routeTemplate.Length + parameters.Length * 8);
-        var pos = 0;
-        var paramIndex = 0;
-
-        while (pos < routeTemplate.Length)
-        {
-            var open = routeTemplate.IndexOf('{', pos);
-            if (open < 0)
-            {
-                // No more placeholders
-                sb.Append(routeTemplate.AsSpan(pos));
-                break;
-            }
-
-            // Append literal text before the placeholder
-            sb.Append(routeTemplate.AsSpan(pos, open - pos));
-
-            var close = routeTemplate.IndexOf('}', open + 1);
-            if (close < 0)
-            {
-                // Malformed template; append rest and break
-                sb.Append(routeTemplate.AsSpan(open));
-                break;
-            }
-
-            // If no more parameter values provided, leave placeholder as-is
-            if (paramIndex >= parameters.Length)
-            {
-                sb.Append(routeTemplate.AsSpan(open, close - open + 1));
-            }
-            else
-            {
-                var parameterValue = parameters[paramIndex]?.ToString() ?? string.Empty;
-                sb.Append(parameterValue);
-                paramIndex++;
-            }
-
-            pos = close + 1;
-        }
-
-        return sb.ToString();
-    }
+    
 }
